refactor: split GhosttySurfaceView surface-I/O methods into a +IO extension
GhosttySurfaceView.swift had grown to the 1000-line cap. moved the surface I/O and binding-action methods
(type/paste into the pty, read selection/buffer/foreground pid, font-size + search binding actions, and the
live font-size read) into GhosttySurfaceView+IO.swift, matching the existing +Config/+Input/+Tracking/+Panes
split. Methods only, no logic change; the main file drops to 839 lines.
